Partial updates to plugin system and added Scene
